+++ src/Symfony/Component/ObjectMapper/Tests/Fixtures/DepthProxy/PostDto.php
@@ -1,0 +1,18 @@
+<?php
+
+namespace Symfony\Component\ObjectMapper\Tests\Fixtures\DepthProxy;
+
+/*
+ * This file is part of the Symfony package.
+ *
+ * (c) Fabien Potencier <fabien@symfony.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+class PostDto
+{
+    public string $title;
+    public ?object $comments = null;
+}